% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kv}
\alias{get_kv}
\title{Get a single key/value from Consul}
\usage{
get_kv(key)
}
\arguments{
\item{key}{which key to retrieve}
}
\value{
Decoded value retrieved from Consul. 'NA' if empty.
}
\description{
Get a single key/value from Consul
}
\examples{
\dontrun{
get_kv("fruit/apple")
}

}
