% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_script}
\alias{init_script}
\alias{init_config}
\alias{init_plumber}
\title{Initiate Commonly Used Script}
\usage{
init_script(name, path = ".", overwrite = FALSE)

init_config(path = ".", overwrite = FALSE)

init_plumber(path = ".", overwrite = FALSE)
}
\arguments{
\item{name}{script selection}

\item{path}{Path to generate template. Default being the current directory.}

\item{overwrite}{Overwrite? Default = FALSE}
}
\value{
Boolean - True if success, False if failure
}
\description{
Initiate commonly used script:

\code{init_config} - To declare database and API configuration

\code{init_plumber} - To convert R code to web API

Please modify content accordingly.
}
