% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_mongo_conn}
\alias{est_mongo_conn}
\title{Establish connection to MongoDB}
\usage{
est_mongo_conn(mongo_id)
}
\arguments{
\item{mysql_id}{Database to connect to}
}
\value{
Database connection if successful.
}
\description{
Fetch credentials from Consul k/v store and establish connection to MySQL.
Required values from k/v store are username, password, host, port, database, collection.
}
\examples{
\dontrun{
est_mongo_conn('somedb')
}
}
