% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_mysql_conn}
\alias{est_mysql_conn}
\title{Establish connection to MySQL database}
\usage{
est_mysql_conn(mysql_id, drv = MySQL())
}
\arguments{
\item{mysql_id}{Database to connect to}

\item{drv}{MySQL driver. Default RMySQL::MySQL}
}
\value{
Database connection if successful.
}
\description{
Fetch credentials from Consul k/v store and establish connection to MySQL.
Required values from k/v store are username, password, host, port, database
}
\examples{
\dontrun{
est_mysql_conn('somedb')
}
}
