% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kv}
\alias{get_kv}
\alias{get_batch_kv}
\title{Obtain Key / Value from Consul}
\usage{
get_kv(key)

get_batch_kv(path, ...)
}
\arguments{
\item{key}{Key name to retrieve}

\item{path}{Path name to retrieve}

\item{...}{One or many keys under path to retrieve.
All but path name will be coersed into a list.}
}
\value{
Decoded value retrieved from Consul. 'NA' if empty.
}
\description{
See https://www.consul.io/ for more about Consul.
}
\examples{
\dontrun{

# Return a string
get_kv("fruit/apple")

# Return a list of values
get_batch_kv("fruit", c("apple", "banana", "coconut"))
# This works too
get_batch_kv("fruit", "apple", "banana", "coconut")

}

}
