% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_config}
\alias{init_config}
\title{Create template for 'Config.R'}
\usage{
init_config(path)
}
\arguments{
\item{path}{Path to generate template}
}
\value{
Boolean - True if success, False if failure
}
\description{
Create template for 'Config.R'
}
